PrinProg Python Project
This Project uses regular expressions to search strings of data and finds lines which have a username and the number of pages they've printed. If the user exisits in the dictionary the number of pages they've printed is added to their value if the user does not exist a new key-value pair is created associating their username with their pages.